fix(cardinal): remove leading slash from endpoint
